[round-3] 상품 조회 / 좋아요 등록 / 주문 / 결제 API 구현

+++ apps/commerce-api/src/main/kotlin/com/loopers/interfaces/api/point/PointChargeRequest.kt
@@ -6,7 +6,7 @@ import jakarta.validation.constraints.Min
 
 data class PointChargeRequest(
     @field:Min(value = 1, message = "충전 금액은 1원 이상이어야 합니다.")
-    val point: Int?,
+    val point: Long?,
 ) {
     fun toCommand(loginId: String): ChargePointCommand = ChargePointCommand(
         loginId = LoginId(loginId),
