+++ apps/commerce-api/src/main/kotlin/com/loopers/interfaces/api/point/PointResponse.kt
@@ -3,7 +3,7 @@ package com.loopers.interfaces.api.point
 import com.loopers.application.point.PointOutput
 
 data class PointResponse(
-    val point: Int,
+    val point: Long,
 ) {
     companion object {
         fun from(output: PointOutput): PointResponse = PointResponse(output.point)
